feat(demo_v7): ST3215 迁移至 FSUS (HX8-U26H-M) 总线舵机协议
协议层:
- 新建 drv_fsus.c/h: Fashion Star UART Servo 协议编解码
- 帧头 0x4C12/0x1C05, checksum = sum & 0xFF (非取反)
- 实现 Ping / SetAngleByVelocity / QueryAngle / ServoMonitor / Stop
- 新建 drv_h13.c/h: HC-13 蓝牙透传驱动(未接线,预置)
- 修复 checksum 算法: ~sum → sum (P0 致命 bug,原因帧被舵机拒绝)

硬件层:
- USART2 从半双工 DMA 改为全双工中断 RX + 阻塞 TX
- bsp_uart 新增 FSUS API: Send / Recv / Flush
- USART2 波特率 1M → 115200
- CMakeLists 修复: CubeMX 覆盖后补回 User include + GLOB + EXCLUDE

任务层:
- task_motion.h/c 重写: MotionCmd/MotionState 适配 FSUS 类型
- 角度映射 0..359 → -180..+180 (map_angle_to_fsus)
- 收包帧同步: 按 0x05 0x1C 头 + size 字段收完整帧
- 修复帧长计算: 多算 1 字节导致 ServoMonitor 超时
- task_manipulator/task_arbiter 字段名适配
- app_control force 模式适配新 MotionCmd

+++ User/task/Inc/task_manipulator.h
@@ -35,7 +35,7 @@ typedef enum {
 typedef struct {
     ManipulatorState_t current_state;
     uint32_t           state_enter_ms;
-    int16_t            last_target_angle_deg;
+    float              last_target_angle_deg;
 } TaskManipulator_Context_t;
 
 /* ============================================================================
